Create DNS-Verify-Fast.ps1
This script does not evaluate whether or not any device is reachable or even real.  It evaluates if FWD DNS entries match PTR DNS entries.  Those that don't have a FWD record that matches their PTR record are output to screen in red - those that just don't match (which usually includes duplicates) are output to screen as normal text.

I've attempted to comment thoroughly enough that this can be adjusted as needed.